Authentication process moved to auth.js
