+++ src/clients/edit/edit.html
@@ -22,18 +22,7 @@
   <!-- ------------------------------------------------------------------------ -->
   <!-- Nav bar -->
 
-  <div class="NavBar ToolBar">
-    <img class="logo" src="/common/images/Melinda-logo-white.png"/>
-    <div class="AppName">Muokkaus</div>
-    <!--
-    <button onClick='reload()'>Reload</button>
-    <button onClick='showTab("loading")'>Loading</button>
-    <button onClick='showTab("login")'>Login</button>
-    <button onClick='showTab("muuntaja")'>Muuntaja</button>
-    <!-- -->
-  </div>
-
-  <div id="progressbar" class="progress-bar-inactive"></div>
+  <div id="navbar"></div>
 
   <!-- ------------------------------------------------------------------------ -->
   <!-- Application "tabs" -->
@@ -94,16 +83,25 @@
       </div>
       <div class="record-merge-panel">
 
-        <div id="record1" class="col BoxShadow">
+        <div id="source" class="col BoxShadow">
           <form class="ToolBar HBox header">
             <div class="Filler VBox Input">
-              <label>Tietue 1</label>
+              <label>Alkuperäinen tietue</label>
               <input id="ID"/>
               </div>
             <button class="material-icons" onclick="return doFetch(event)">search</button>
             </form>
           <div id="Record" class="marc-record">
           </div>
+        </div>
+
+        <div id="result" class="col BoxShadow">
+          <div class="ToolBar HBox header">
+            <div class="Filler">TULOSTIETUE</div>
+            <button class="material-icons" onClick="onSave(event)" disabled>save</button>
+          </div>
+          <div id="Record" class="marc-record">
+            </div>
         </div>
 
       </div>
