+++ src/clients/edit/edit.js
@@ -8,9 +8,10 @@ import {startProcess, stopProcess} from "/common/ui-utils.js";
 import {showTab, resetForms, reload} from "/common/ui-utils.js";
 import {createMenuBreak, createMenuItem, createMenuSelection} from "../common/ui-utils.js";
 
-import {Account} from "/common/auth.js"
+import {Account, doLogin, logout} from "/common/auth.js"
 import {transformRequest} from "/common/rest.js";
 import {showRecord} from "/common/marc-record-ui.js";
+import {setNavBar} from "../common/navbar.js";
 
 var transformed = {
   source: {},
@@ -24,63 +25,16 @@ var transformed = {
 window.initialize = function () {
   console.log('Initializing');
 
-  Account.verify()
-    .then(response => authSuccess(Account.get()))
-    .catch(noAuth);
+  setNavBar(document.querySelector('#navbar'), "Muokkaus")
 
-  function noAuth() {
-    Account.remove();
-    resetForms(document.getElementById('root'));
-    showTab('login');
+  doLogin(authSuccess);
+
+  function authSuccess(user) {
+    const username = document.querySelector("#account-menu #username")
+    username.innerHTML = Account.get()["Name"];
+    showTab('muuntaja');
+    doFetch();
   }
-}
-
-//-----------------------------------------------------------------------------
-// Login & logout
-//-----------------------------------------------------------------------------
-
-window.login = function (e) {
-  e.preventDefault();
-
-  console.log('Login:', e);
-
-  logininfo('');
-
-  const termschecked = document.querySelector('#login #acceptterms').checked;
-  if (!termschecked) {
-    logininfo('Tietosuojaselosteen hyväksyminen vaaditaan');
-    return;
-  }
-
-  const username = document.getElementById('username').value;
-  const password = document.getElementById('password').value;
-
-  startProcess();
-
-  Account.login(username, password)
-    .then(user => authSuccess(user))
-    .catch(err => {
-      Account.remove();
-      logininfo('Tunnus tai salasana ei täsmää');
-    })
-    .finally(stopProcess);
-
-  function logininfo(msg) {
-    const infodiv = document.querySelector('#login #info');
-    infodiv.innerHTML = msg;
-  }
-}
-
-function logout(e) {
-  Account.logout();
-  reload();
-}
-
-function authSuccess(user) {
-  const username = document.querySelector("#account-menu #username")
-  username.innerHTML = Account.get()["Name"];
-  showTab('muuntaja');
-  doFetch();
 }
 
 //-----------------------------------------------------------------------------
@@ -109,7 +63,7 @@ window.doFetch = function (event = undefined) {
   console.log('Fetching...');
   if (event) event.preventDefault();
 
-  const sourceID = document.querySelector(`#muuntaja .record-merge-panel #record1 #ID`).value;
+  const sourceID = document.querySelector(`#muuntaja .record-merge-panel #source #ID`).value;
   //const baseID = document.querySelector(`#muuntaja .record-merge-panel #record2 #ID`).value;
 
   if (!transformed.source || sourceID != transformed.source.ID) {
@@ -127,6 +81,7 @@ window.doFetch = function (event = undefined) {
     .then(records => {
       stopProcess();
       console.log('Fetched:', records);
-      showRecord(records.source, 'record1');
+      showRecord(records.source, 'source');
+      showRecord(records.source, 'result');
     });
 }
