+++ src/clients/muuntaja/muuntaja.html
@@ -25,18 +25,7 @@
   <!-- ------------------------------------------------------------------------ -->
   <!-- Nav bar -->
 
-  <div class="NavBar ToolBar">
-    <img class="logo" src="/common/images/Melinda-logo-white.png"/>
-    <div class="AppName">Muuntaja</div>
-    <!--
-    <button onClick='reload()'>Reload</button>
-    <button onClick='showTab("loading")'>Loading</button>
-    <button onClick='showTab("login")'>Login</button>
-    <button onClick='showTab("muuntaja")'>Muuntaja</button>
-    <!-- -->
-  </div>
-
-  <div id="progressbar" class="progress-bar-inactive"></div>
+  <div id="navbar"></div>
 
   <!-- ------------------------------------------------------------------------ -->
   <!-- Application "tabs" -->
